Excel file used to record IOTA address growth
Address growth gives an indication of where the network sits on the applicable Metcalfe's Netoid, which describes the overall growth of the network, and subsequently, the value of the network from Metcalfe's law, V = p*N^2, where p is a measure of the network utility. This is the only factor which value depends upon. With no change in p, the rate at which growth proceeds will increase towards a maximum rate of change around halfway through the lifetime of the network entire expansion. Note that the utility of Bitcoin and other Blockchain based distributed ledger solutions dropped significantly when scaling issues were encountered in those networks. By Metcalfe's law, that change of utility actually happened very early in the potential growth of Bitcoin. $20,000 per bitcoin, the maximum price that the network token achieved towards the end of 2018 when said disaster struck, is nothing to what it would have continued towards, if that problem had not occurred. Also interestingly, Note also Ethereum experienced a much higher growth rate than Bitcoin. This was because Ethereum had a higher utility than Bitcoin with smart contracts etc. Subsequently, Ethereum hit the scaleability wall in a much shorter time than Bitcoin. IOTA now uniquely stands as the next contender, as it the only solution with no such scaling limitations. Analyse the address growth, and record here to eliminate any doubts.